pyproject.toml
[project]
name = "mistral-ner"
version = "0.2.0"
description = "Fine-tune Mistral-7B for Named Entity Recognition (NER) on the CoNLL-2003 dataset using Hugging Face Transformers."
authors = [
    {name = "Sergei Nevedomski", email = "neviadomski@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "accelerate>=1.7.0",
    "datasets>=3.6.0",
    "evaluate>=0.4.3",
    "numpy>=2.2.5",
    "protobuf>=6.31.0",
    "peft>=0.4.0",
    "sentencepiece>=0.2.0",
    "seqeval>=1.2.2",
    "transformers>=4.51.3",
    "wandb>=0.16.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
cuda = [
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "bitsandbytes>=0.41.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
cuda11 = [
    "torch>=2.7.0+cu118",
    "torchaudio>=2.7.0+cu118",
    "torchvision>=0.22.0+cu118",
    "bitsandbytes>=0.41.0",
]
cuda12 = [
    "torch>=2.7.0+cu121",
    "torchaudio>=2.7.0+cu121",
    "torchvision>=0.22.0+cu121",
    "bitsandbytes>=0.41.0",
]

[project.urls]
Homepage = "https://github.com/nevedomski/mistral-ner"
Documentation = "https://github.com/nevedomski/mistral-ner#readme"

[tool.uv]
index-url = "https://pypi.org/simple"
extra-index-url = ["https://download.pytorch.org/whl/cu121", "https://download.pytorch.org/whl/cu118"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "extra == 'cuda12'" },
    { index = "pytorch-cu118", marker = "extra == 'cuda11'" }
]
torchaudio = [
    { index = "pytorch-cu121", marker = "extra == 'cuda12'" },
    { index = "pytorch-cu118", marker = "extra == 'cuda11'" }
]
torchvision = [
    { index = "pytorch-cu121", marker = "extra == 'cuda12'" },
    { index = "pytorch-cu118", marker = "extra == 'cuda11'" }
]
